20161003 - confirm delete

+++ resources/views/admin/permissions/show.blade.php
@@ -47,7 +47,7 @@
        
             <div class='row'>
                 <div class='col-md-1'><a href="{{ action('PermissionsController@edit', $permission->id) }}" class="btn btn-info">{!! Html::image('img/edit.png', 'Edit',array('title'=>"Edit")) !!}</a></div>
-                <div class='col-md-1'>{!! Form::open(['method' => 'DELETE', 'route' => ['admin.permission.destroy', $permission->id]]) !!}
+                <div class='col-md-1'>{!! Form::open(['method' => 'DELETE', 'route' => ['admin.permission.destroy', $permission->id],'onsubmit'=>'return ConfirmDelete()']) !!}
                 {!! Form::image('img/delete.png','btnDelete',['class' => 'btn btn-danger','title'=>'Delete']) !!} 
                 {!! Form::close() !!}</div><div class="clearfix"> </div>
             </div>
