+++ resources/views/contacts/show.blade.php
@@ -137,7 +137,7 @@
         </div>
         <div class='row'>
             <div class='col-md-1'><a href="{{ action('PersonsController@edit', $contact->id) }}" class="btn btn-info">{!! Html::image('img/edit.png', 'Edit',array('title'=>"Edit")) !!}</a></div>
-            <div class='col-md-1'>{!! Form::open(['method' => 'DELETE', 'route' => ['person.destroy', $contact->id]]) !!}
+            <div class='col-md-1'>{!! Form::open(['method' => 'DELETE', 'route' => ['person.destroy', $contact->id],'onsubmit'=>'return ConfirmDelete()']) !!}
             {!! Form::image('img/delete.png','btnDelete',['class' => 'btn btn-danger','title'=>'Delete']) !!} 
             {!! Form::close() !!}</div><div class="clearfix"> </div>
         </div>
