+++ resources/views/dioceses/show.blade.php
@@ -198,7 +198,7 @@
             </div>            
                 <div class='row'>
                     <div class='col-md-1'><a href="{{ action('DiocesesController@edit', $diocese->id) }}" class="btn btn-info">{!! Html::image('img/edit.png', 'Edit',array('title'=>"Edit")) !!}</a></div>
-                    <div class='col-md-1'>{!! Form::open(['method' => 'DELETE', 'route' => ['diocese.destroy', $diocese->id]]) !!}
+                    <div class='col-md-1'>{!! Form::open(['method' => 'DELETE', 'route' => ['diocese.destroy', $diocese->id],'onsubmit'=>'return ConfirmDelete()']) !!}
                     {!! Form::image('img/delete.png','btnDelete',['class' => 'btn btn-danger','title'=>'Delete']) !!} 
                     {!! Form::close() !!}</div><div class="clearfix"> </div>
                 </div>
