+++ resources/views/rooms/show.blade.php
@@ -28,7 +28,7 @@
                 </div><div class="clearfix"> </div>
                 <div class='row'>
                     <div class='col-md-1'><a href="{{ action('RoomsController@edit', $room->id) }}" class="btn btn-info">{!! Html::image('img/edit.png', 'Edit',array('title'=>"Edit")) !!}</a></div>
-                    <div class='col-md-1'>{!! Form::open(['method' => 'DELETE', 'route' => ['room.destroy', $room->id]]) !!}
+                    <div class='col-md-1'>{!! Form::open(['method' => 'DELETE', 'route' => ['room.destroy', $room->id],'onsubmit'=>'return ConfirmDelete()']) !!}
                     {!! Form::image('img/delete.png','btnDelete',['class' => 'btn btn-danger','title'=>'Delete']) !!} 
                     {!! Form::close() !!}</div><div class="clearfix"> </div>
                 </div>
